assignment2-codebase-genius/requirements.txt
streamlit
requests
markdown
byllm
jac-cloud
GitPython
networkx
graphviz
jinja2
tree-sitter
matplotlib
